refactor : 핸들러, 옵션 분리
