+++ src/components/admin/management/cocktailItemOptions.ts
@@ -1,0 +1,7 @@
+import { CocktailProps } from "../../../libs/interface/interfaceBarDetail";
+
+export interface CocktailItemProps {
+  id: string;
+  setShowList: Function;
+  info: CocktailProps;
+}
